+++ src/main/java/com/pax/market/api/sdk/java/api/emm/emmDeviceDetail/dto/EmmDeviceLocationResponse.java
@@ -1,0 +1,9 @@
+package com.pax.market.api.sdk.java.api.emm.emmDeviceDetail.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.Response;
+
+public class EmmDeviceLocationResponse  extends Response<EmmDeviceLocationDTO> {
+
+    private static final long serialVersionUID = 8691575753427719275L;
+
+}